LPS-200190 Move logic to BNDBreakingChangeCommitMessageCheck.java
